working on epsilon states
