+++ epsilon.ndfa
@@ -1,0 +1,26 @@
+start = q0
+accept = [q6]
+
+q0:
+    0   q1
+    eps q2
+q1:
+    $log (in q1)
+    0   q3
+    1   q4
+q2:
+    $log (in q2 (epsilon branch))
+    0   q5
+    1   q6
+q3:
+    0   null
+    1   null
+q4:
+    0   null
+    1   null
+q5:
+    0   null
+    1   null
+q6:
+    0   null
+    1   null
